refactor: remove infromation popover [SF-618]
