--- src/components/atoms/InformationPopover/index.ts
@@ -1,1 +1,0 @@
-export * from './InformationPopover';
