+++ src/components/atoms/OrderDisplayBox/OrderDisplayBox.tsx
@@ -1,4 +1,4 @@
-import { InformationPopover } from 'src/components/atoms';
+import { Tooltip } from 'src/components/templates';
 
 interface OrderDisplayBoxProps {
     field: string;
@@ -15,9 +15,7 @@ export const OrderDisplayBox = ({
         <div className='typography-caption flex h-6 w-full flex-row items-center justify-between bg-transparent'>
             <div className='flex flex-row items-center gap-2'>
                 <div className='text-slateGray'>{field}</div>
-                {informationText && (
-                    <InformationPopover>{informationText}</InformationPopover>
-                )}
+                {informationText && <Tooltip>{informationText}</Tooltip>}
             </div>
             <div className='text-right text-planetaryPurple'>{value}</div>
         </div>
