+++ src/components/molecules/CollateralProgressBar/CollateralProgressBar.test.tsx
@@ -38,9 +38,9 @@ describe('CollateralProgressBar Component', () => {
         const information = screen.getByTestId('information-circle');
         fireEvent.mouseEnter(information);
 
-        const informationPopover = screen.getByTestId('information-popover');
+        const tooltip = screen.getByTestId('tooltip');
 
-        expect(informationPopover).toHaveTextContent(
+        expect(tooltip).toHaveTextContent(
             'Your total borrow limit is at $80.00 which is 80% of your $100.00 collateral deposit.Increasing collateral deposit will increase your borrow limit by 80% of its value.'
         );
     });
