+++ src/components/molecules/LiquidationProgressBar/LiquidationProgressBar.test.tsx
@@ -35,9 +35,9 @@ describe('LiquidationProgressBar Component', () => {
         const information = screen.getByTestId('information-circle');
         fireEvent.mouseEnter(information);
 
-        const informationPopover = screen.getByTestId('information-popover');
+        const tooltip = screen.getByTestId('tooltip');
 
-        expect(informationPopover).toHaveTextContent(
+        expect(tooltip).toHaveTextContent(
             'Liquidation threshold is the limit where your collateral will be eligible for liquidation.You are currently 35% under the liquidation threshold (80% of deposit balance).'
         );
     });
