+++ src/components/organisms/UnwindDialog/UnwindDialog.tsx
@@ -5,7 +5,6 @@ import { useCallback, useMemo, useReducer, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
     ExpandIndicator,
-    InformationPopover,
     Section,
     SectionWithItems,
     Spinner,
@@ -18,6 +17,7 @@ import {
     FailurePanel,
     SuccessPanel,
 } from 'src/components/molecules';
+import { Tooltip } from 'src/components/templates';
 import {
     useCollateralBook,
     useEtherscanUrl,
@@ -306,10 +306,10 @@ const feeItem = () => {
     return (
         <div className='flex flex-row items-center gap-1'>
             <div className='text-planetaryPurple'>Transaction Fee %</div>
-            <InformationPopover>
+            <Tooltip>
                 A duration-based transaction fee only for market takers,
                 factored into the bond price, and deducted from its future value
-            </InformationPopover>
+            </Tooltip>
         </div>
     );
 };
