+++ src/components/templates/Tooltip/Tooltip.test.tsx
@@ -11,17 +11,17 @@ describe('Tooltip Component', () => {
 
     it('should render information on mouseEnter event', () => {
         render(<Default />);
-        const information = screen.getByTestId('button-icon');
+        const information = screen.getByTestId('information-circle');
         fireEvent.mouseEnter(information);
 
-        expect(screen.getByTestId('information-popover')).toBeInTheDocument();
+        expect(screen.getByTestId('tooltip')).toBeInTheDocument();
         expect(screen.getByText('Tooltip content')).toBeInTheDocument();
         expect(screen.getByRole('tooltip')).toBeInTheDocument();
     });
 
     it('should remove rendered information on mouseOut event', () => {
         render(<Default />);
-        const information = screen.getByTestId('button-icon');
+        const information = screen.getByTestId('information-circle');
         fireEvent.mouseEnter(information);
         fireEvent.mouseOut(information);
